[sdk][typescript] Refactor bcs
