+++ ecosystem/typescript/sdk/.eslintignore
@@ -2,4 +2,3 @@ node_modules
 packages/**/node_modules/
 dist/**
 **/*.test.ts
-src/transaction_builder/**
