+++ ecosystem/typescript/sdk/.eslintrc.js
@@ -4,7 +4,7 @@ module.exports = {
     es2021: true,
     node: true,
   },
-  extends: ["airbnb-base"],
+  extends: ["airbnb-typescript/base"],
   parser: "@typescript-eslint/parser",
   parserOptions: {
     ecmaVersion: "latest",
