+++ ecosystem/typescript/sdk/src/transaction_builder/aptos_types/identifier.ts
@@ -1,0 +1,14 @@
+import { Deserializer, Serializer } from "../bcs";
+
+export class Identifier {
+  constructor(public value: string) {}
+
+  public serialize(serializer: Serializer): void {
+    serializer.serializeStr(this.value);
+  }
+
+  static deserialize(deserializer: Deserializer): Identifier {
+    const value = deserializer.deserializeStr();
+    return new Identifier(value);
+  }
+}
