+++ ecosystem/typescript/sdk/src/transaction_builder/aptos_types/index.ts
@@ -1,0 +1,5 @@
+export * from "./account_address";
+export * from "./authenticator";
+export * from "./transaction";
+export * from "./type_tag";
+export * from "./identifier";
