+++ ecosystem/typescript/sdk/src/transaction_builder/bcs/types.ts
@@ -1,0 +1,10 @@
+export type Seq<T> = T[];
+
+export type unit = null;
+export type bool = boolean;
+export type uint8 = number;
+export type uint16 = number;
+export type uint32 = number;
+export type uint64 = bigint;
+export type uint128 = bigint;
+export type bytes = Uint8Array;
