--- ecosystem/typescript/sdk/src/transaction_builder/serde/index.ts
@@ -1,5 +1,0 @@
-export * from "./types";
-export * from "./serializer";
-export * from "./deserializer";
-export * from "./binarySerializer";
-export * from "./binaryDeserializer";
